requirements.txt
gradio
google-generativeai
pandas
pillow
